healthelife/building_pagelets_2019/requirements.txt
flask
requests
python-jose
fhirclient == 1.0.3
pyopenssl